files_to_text/requirements.txt
fastapi 
uvicorn 
pillow 
pdf2image 
requests